Create base class for displaying collections
